add file icon to editor tabs
