+++ ui/src/ide/tabs/TabIndex.scss
@@ -1,0 +1,4 @@
+.tabIcon {
+  height: 18px;
+  margin-right: 6px;
+}
